cleaned up runregistry in the same way as runnumber
